Array DS
Using two pointers technique.
This program prints all the element of an array in reverse order. For simplicity I have taken (as input) and reversed integer type of data. I also used a function here, the only task of this function is to reverse the array. And I have written the program to print the reversed array in the main function.